requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OCR and Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
pdfplumber==0.10.3
opencv-python==4.9.0.80
pillow==10.2.0

# Vector DB and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.13

# Language Processing
transformers==4.37.0
torch==2.1.2
accelerate==0.26.1
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
pandas==2.0.3